+++ utils/vec4.cpp
@@ -8,7 +8,7 @@ vec4::vec4()
 	x = y = z = w = 0;
 }
 
-vec4::vec4(const vec3 &vec, double w) 
+vec4::vec4(const vec3 &vec, float w) 
 {
 	this->x = vec.x;
 	this->y = vec.y;
@@ -16,7 +16,7 @@ vec4::vec4(const vec3 &vec, double w)
 	this->w = w;
 }
 
-vec4::vec4(double x, double y, double z, double w) 
+vec4::vec4(float x, float y, float z, float w) 
 {
 	this->x = x;
 	this->y = y;
@@ -24,7 +24,7 @@ vec4::vec4(double x, double y, double z, double w)
 	this->w = w;
 }
 
-vec4& vec4::set(double x, double y, double z, double w) {
+vec4& vec4::set(float x, float y, float z, float w) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
@@ -55,12 +55,12 @@ vec4 vec4::operator-(const vec4& v) const {
 	return vec4(x - v.x, y - v.y, z - v.z, w - v.w);
 }
 
-vec4 vec4::operator*(double scalar) const {
+vec4 vec4::operator*(float scalar) const {
 	return vec4(x * scalar, y * scalar, z * scalar, w * scalar);
 }
 
-vec4 vec4::operator/(double scalar) const {
-	double inv = 1.0f / scalar;
+vec4 vec4::operator/(float scalar) const {
+	float inv = 1.0f / scalar;
 	return vec4(x * inv, y * inv, z * inv, w * inv);
 }
 
@@ -88,7 +88,7 @@ vec4& vec4::operator-=(const vec4& v) {
 	return *this;
 }
 
-vec4& vec4::operator*=(double scalar) {
+vec4& vec4::operator*=(float scalar) {
 	x *= scalar;
 	y *= scalar;
 	z *= scalar;
@@ -96,8 +96,8 @@ vec4& vec4::operator*=(double scalar) {
 	return *this;
 }
 
-vec4& vec4::operator/=(double scalar) {
-	double inv = 1.0f / scalar;
+vec4& vec4::operator/=(float scalar) {
+	float inv = 1.0f / scalar;
 	x *= inv;
 	y *= inv;
 	z *= inv;
@@ -105,32 +105,32 @@ vec4& vec4::operator/=(double scalar) {
 	return *this;
 }
 
-double& vec4::operator[](int index) {
+float& vec4::operator[](int index) {
 	assert(index >= 0 && index < 4);
 	return (&x)[index];
 }
 
-const double& vec4::operator[](int index) const {
+const float& vec4::operator[](int index) const {
 	assert(index >= 0 && index < 4);
 	return (&x)[index];
 }
 
-double vec4::dot(const vec4& v) const {
+float vec4::dot(const vec4& v) const {
 	return x * v.x + y * v.y + z * v.z + w * v.w;
 }
 
-double vec4::magnitude(void) const {
+float vec4::magnitude(void) const {
 	return sqrtf(x * x + y * y + z * z + w * w);
 }
 
 vec4& vec4::normalize(void) {
-	double mag = sqrtf(x * x + y * y + z * z + w * w);
+	float mag = sqrtf(x * x + y * y + z * z + w * w);
 
 	if (mag < 1e-6f) {
 		x = y = z = w = 0;
 	}
 	else {
-		double inv = 1.0f / mag;
+		float inv = 1.0f / mag;
 
 		x *= inv;
 		y *= inv;
@@ -141,23 +141,23 @@ vec4& vec4::normalize(void) {
 	return *this;
 }
 
-double* vec4::ptr(void) {
+float* vec4::ptr(void) {
 	return &x;
 }
 
-const double* vec4::ptr(void) const {
+const float* vec4::ptr(void) const {
 	return &x;
 }
 
-vec4 operator*(double scalar, const vec4& v) {
+vec4 operator*(float scalar, const vec4& v) {
     printf("x: %f y: %f z: %f w: %f\n", v.x * scalar, v.y * scalar, v.z * scalar, v.w * scalar);
 	return vec4(v.x * scalar, v.y * scalar, v.z * scalar, v.w * scalar);
 }
 
 vec4 operator*(const mat4& mat, const vec4& v) {
-	double x = mat[0] * v.x + mat[1] * v.y + mat[2] * v.z + mat[3] * v.w;
-	double y = mat[4] * v.x + mat[5] * v.y + mat[6] * v.z + mat[7] * v.w;
-	double z = mat[8] * v.x + mat[9] * v.y + mat[10] * v.z + mat[11] * v.w;
-	double w = mat[12] * v.x + mat[13] * v.y + mat[14] * v.z + mat[15] * v.w;
+	float x = mat[0] * v.x + mat[1] * v.y + mat[2] * v.z + mat[3] * v.w;
+	float y = mat[4] * v.x + mat[5] * v.y + mat[6] * v.z + mat[7] * v.w;
+	float z = mat[8] * v.x + mat[9] * v.y + mat[10] * v.z + mat[11] * v.w;
+	float w = mat[12] * v.x + mat[13] * v.y + mat[14] * v.z + mat[15] * v.w;
 	return vec4(x, y, z, w);
 }
